requirements.txt
absl-py==0.11.0
appdirs==1.4.4
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.6
djangorestframework==3.12.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
imageio==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.4
networkx==2.5
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
Pillow==8.1.0
pooch==1.3.0
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.17.1
scipy==1.4.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tifffile==2020.9.3
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
